% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{compute_number_jumps}
\alias{compute_number_jumps}
\title{Compute the number of jumps}
\usage{
compute_number_jumps(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, \code{time} and \code{state} (see \code{\link{generate_Markov_cfd}})}
}
\value{
A vector containing the number of jumps for each individual#'
}
\description{
Compute the number of jumps
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

nJump <- compute_number_jumps(d_JK)

}
\seealso{
\link{hist.njump}
}
\author{
Cristian Preda
}
