% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{compute_optimal_encoding}
\alias{compute_optimal_encoding}
\title{Compute the optimal encodings f each state}
\usage{
compute_optimal_encoding(data_msm, basisobj)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, \code{time} and \code{state} (see \code{\link{generate_Markov_cfd}}). All individual must end at the same time Tmax (use \code{\link{msm2msmTmax}}).}

\item{basisobj}{is a basis create using \code{fda} package}
}
\value{
a list containing:
\itemize{
  \item eigenvalues eigenvalues
  \item alpha optimal encoding coefficients
  \item pc principal components
  \item F 
  \item G  
  \item V 
}
}
\description{
Compute the optimal encodings f each state
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
Tmax <- 10
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = Tmax)
d_JK2 <- msm2msmTmax(d_JK, 10)

# create basis object
m <- 10
b <- create.bspline.basis(c(0, Tmax), nbasis = m, norder = 4)

# compute encoding
encoding <- compute_optimal_encoding(d_JK2, b)

}
\seealso{
plot.fmca
}
\author{
Cristian Preda
}
