% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateEncoding.R
\name{compute_optimal_encoding_multivariate}
\alias{compute_optimal_encoding_multivariate}
\title{Compute the multivariate optimal encoding}
\usage{
compute_optimal_encoding_multivariate(
  data,
  basisobj,
  epsilon = NULL,
  stateColumns = NULL,
  verbose = TRUE,
  nCores = max(1, detectCores() - 1)
)
}
\arguments{
\item{data}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and
and several columns for the associated state for each dimension.
All individuals must begin at the same time T0 and end at the same time Tmax.}

\item{basisobj}{basis created using the \code{fda} package (cf. \code{\link{create.basis}}).
The same basis is used for every dimension.}

\item{epsilon}{epsilon added to the diagonal of F in order to invert it. If NULL, an epsilon is computed with regards to F.
It can be a vector to test several values.}

\item{stateColumns}{column names for multivariate states. By default, "state1", "state2", ...}

\item{verbose}{if TRUE print some information}

\item{nCores}{number of cores used for parallelization. Default is all cores except 1.}
}
\value{
a fmca object
}
\description{
Compute the multivariate optimal encoding
}
\examples{
\donttest{
set.seed(42)
Tmax <- 2
x1 <- generate_Markov(n = 500, K = 2)
x1 <- cut_data(x1, Tmax = Tmax)
x2 <- generate_Markov(n = 500, K = 2)
x2 <- cut_data(x2, Tmax = Tmax)

x <- list(x1, x2)

mvcfd <- convert2mvcfd(x)
basisobj <- create.bspline.basis(c(0, Tmax), nbasis = 10, norder = 4)

multEnc <- compute_optimal_encoding_multivariate(mvcfd, basisobj, verbose = FALSE, nCores = 1)

plot(multEnc)
}
}
