% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{compute_time_spent}
\alias{compute_time_spent}
\title{Compute time spent in each state}
\usage{
compute_time_spent(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}
}
\value{
a matrix with \code{K} columns containing the total time spent in each state for each individuals
}
\description{
For each individual, compute the time spent in each state
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# cut at Tmax = 8
d_JK2 <- cut_data(d_JK, Tmax = 8)

# compute time spent by each id in each state
timeSpent <- compute_time_spent(d_JK2)

}
\seealso{
\link{boxplot.timeSpent}
}
\author{
Cristian Preda, Quentin Grimonprez
}
