% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateFormat.R
\name{convert2mvcfd}
\alias{convert2mvcfd}
\title{Convert a list of cfd into a multivariate cfd}
\usage{
convert2mvcfd(x, stateColumns = NULL)
}
\arguments{
\item{x}{list of cfd}

\item{stateColumns}{column names for multivariate states. By default, "state1", "state2", ...}
}
\value{
a multivariate cfd
}
\description{
Convert a list of cfd into a multivariate cfd
}
\examples{
set.seed(42)
x1 <- generate_Markov(n = 10, K = 2)
x1 <- cut_data(x1, Tmax = 1)
x2 <- generate_Markov(n = 10, K = 2)
x2 <- cut_data(x2, Tmax = 1)

x <- list(x1, x2)

mvcfd <- convert2mvcfd(x)

}
