% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateFormat.R
\name{convertListCfd2Cfd}
\alias{convertListCfd2Cfd}
\title{Convert a list of univariate cfds to an unique univariate cfd}
\usage{
convertListCfd2Cfd(x, sep = "_")
}
\arguments{
\item{x}{list of cfd}

\item{sep}{separator used to concatenate states}
}
\value{
univariate cfd to use in \code{\link{compute_optimal_encoding}}. A data.frame with is, time and state columns
}
\description{
Convert a list of univariate cfds to an unique univariate cfd
}
\examples{
set.seed(42)
x1 <- generate_Markov(n = 10, K = 2)
x1 <- cut_data(x1, Tmax = 1)
x2 <- generate_Markov(n = 10, K = 2)
x2 <- cut_data(x2, Tmax = 1)

x <- list(x1, x2)

cfd <- convertListCfd2Cfd(x)

}
