% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateFormat.R
\name{convertMvcfd2cfd}
\alias{convertMvcfd2cfd}
\title{Convert a multivariate cfd to an univariate one}
\usage{
convertMvcfd2cfd(x, sep = "_", stateColumns = NULL)
}
\arguments{
\item{x}{multivariate cfd: data.frame with id, time columns and several stateColumns (output of \code{\link{convert2mvcfd}})}

\item{sep}{separator used to concatenate states}

\item{stateColumns}{names of the state columns. If NULL, it looks for columns with state in their names}
}
\value{
univariate cfd to use in \code{\link{compute_optimal_encoding}}. A data.frame with is, time and state columns
}
\description{
Convert a multivariate cfd to an univariate one
}
\examples{
set.seed(42)
x1 <- generate_Markov(n = 10, K = 2)
x1 <- cut_data(x1, Tmax = 1)
x2 <- generate_Markov(n = 10, K = 2)
x2 <- cut_data(x2, Tmax = 1)

x <- list(x1, x2)

mvcfd <- convert2mvcfd(x)
cfd <- convertMvcfd2cfd(mvcfd)

}
