% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreatment.R
\name{convertToCfd}
\alias{convertToCfd}
\title{Convert data to categorical functional data}
\usage{
convertToCfd(
  x,
  thr,
  leftClosed = TRUE,
  labels = NULL,
  times = NULL,
  idLabels = NULL,
  nx = 200,
  byrow = FALSE
)
}
\arguments{
\item{x}{matrix or fd object}

\item{thr}{vector defining the intervals to create the categories}

\item{leftClosed}{If \code{TRUE}, intervals generated with \code{thr} are left closed (\code{[thr[i]:thr[i+1])})
otherwise there are right closed (\code{(thr[i]:thr[i+1]]})}

\item{labels}{labels of categories (\code{length(thr) - 1}). \code{abels[i]} is the category
associated with \code{[thr[i]:thr[i+1])}. If \code{NULL}, categories are numbered}

\item{times}{vector containing values at which \code{fd} is to be evaluated}

\item{idLabels}{vector containing id labels. If NULL it use the names found in the matrix or fd object}

\item{nx}{Only if \code{x} is a fd object. Number of points to evaluate \code{fd}}

\item{byrow}{Only if \code{x} is a matrix. If \code{FALSE}, one column = one trajectory}
}
\value{
a data.frame in the cfda format
}
\description{
Convert data to categorical functional data
}
\examples{
# fd object
data("CanadianWeather")
temp <- CanadianWeather$dailyAv[,, "Temperature.C"]
basis <- create.bspline.basis(c(1, 365), nbasis = 8, norder = 4)
fd <- smooth.basis(1:365, temp, basis)$fd

# "Very Cold" = [-50:-10), "Cold" = [-10:0), ...
out <- convertToCfd(fd, thr = c(-50, -10, 0, 10, 20, 50),
                    labels = c("Very Cold", "Cold", "Fresh", "OK", "Hot"),
                    times = 1:365)

# matrix
out2 <- convertToCfd(temp, thr = c(-50, -10, 0, 10, 20, 50),
                     labels = c("Very Cold", "Cold", "Fresh", "OK", "Hot"),
                     times = 1:365, byrow = FALSE)

}
