% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation_Markov}
\alias{estimation_Markov}
\title{Estimate transition matrix and spent time}
\usage{
estimation_Markov(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, \code{time} and \code{state} (see \code{\link{generate_Markov_cfd}})}
}
\value{
list of two elements: \code{Q}, the estimated transition matrix, adn \code{lambda}, the estimated time spent in each state
}
\description{
Estimate transition matrix and spent time
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# estimation  
mark <- estimation_Markov(d_JK)
mark$Q
mark$lambda

}
\seealso{
\link{plot.Markov}
}
\author{
Cristian Preda
}
