% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreatment.R
\name{fdToCfd}
\alias{fdToCfd}
\title{Convert functional data to categorical functional data}
\usage{
fdToCfd(fd, thr, leftClosed = TRUE, labels = NULL, evalarg = NULL, nx = 200)
}
\arguments{
\item{fd}{fd object}

\item{thr}{vector defining the intervals to create the categories}

\item{leftClosed}{If \code{TRUE}, intervals generated with \code{thr} are left closed (\code{[thr[i]:thr[i+1])})
otherwise there are right closed (\code{(thr[i]:thr[i+1]]})}

\item{labels}{labels of categories (\code{length(thr) - 1}). \code{abels[i]} is the category
associated with \code{[thr[i]:thr[i+1])}. If \code{NULL}, categories are numbered}

\item{evalarg}{vector containing values at which \code{fd} is to be evaluated}

\item{nx}{number of points to evaluate \code{fd}}
}
\value{
a data.frame in the cfda format
}
\description{
Convert functional data to categorical functional data
}
\examples{
data("CanadianWeather")
temp <- CanadianWeather$dailyAv[,, "Temperature.C"]
basis <- create.bspline.basis(c(1, 365), nbasis = 8, norder = 4)
fd <- smooth.basis(1:365, temp, basis)$fd

# "Very Cold" = [-50:-10), "Cold" = [-10:0), ...
fdToCfd(fd, thr = c(-50, -10, 0, 10, 20, 50),
        labels = c("Very Cold", "Cold", "Fresh", "OK", "Hot"),
        evalarg = 1:365)

}
