% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate_2State}
\alias{generate_2State}
\title{Generate data following a 2 states model}
\usage{
generate_2State(n)
}
\arguments{
\item{n}{number of individuals}
}
\value{
a data.frame with 3 columns : \code{id}, id of the trajectory, \code{time}, time at chich a change occurs and \code{state}, new state.
}
\description{
Generate data following a 2 states model
}
\details{
Let \eqn{\theta\sim \mathcal{U}[0, 1]}

The state at time t is defined by \eqn{X_t(w) = 1} if \eqn{t < \theta(w)}, 2 otherwise.
}
\author{
Cristian Preda, Quentin Grimonprez
}
