% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate_Markov_cfd}
\alias{generate_Markov_cfd}
\title{Generate Markov Trajectories}
\usage{
generate_Markov_cfd(n = 5, K = 2, Q = 1 - diag(K), lambda = rep(1,
  K), pi_0 = c(1, rep(0, K - 1)), Tmax = 1)
}
\arguments{
\item{n}{number of trajectories to generate}

\item{K}{number of states}

\item{Q}{matrix which indicates the allowed transitions in the continuous-time Markov chain, and optionally also the initial values of those transitions}

\item{lambda}{time spent in each state}

\item{pi_0}{initial distribution of states}

\item{Tmax}{maximal duration of trajectories}
}
\value{
a data.frame with 3 columns : \code{id}, id of the trajectory, \code{time}, time at chich a change occurs and \code{state}, new state.
}
\description{
Generate Markov Trajectories
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK = generate_Markov_cfd(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

head(d_JK)


}
\author{
Cristian Preda
}
