% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{hist.njump}
\alias{hist.njump}
\title{Plot the number of jumps}
\usage{
\method{hist}{njump}(x, breaks = NULL, ...)
}
\arguments{
\item{x}{output of \code{\link{compute_number_jumps}} function}

\item{breaks}{number of breaks. If not given use the sturges rule}

\item{...}{not used}
}
\description{
Plot the number of jumps
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

nJump <- compute_number_jumps(d_JK)

hist(nJump)

}
\author{
Quentin Grimonprez
}
