% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{plot.Markov}
\alias{plot.Markov}
\title{Plot the transition graph}
\usage{
\method{plot}{Markov}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{estimation_Markov}} function}

\item{...}{not used}
}
\description{
Plot the transition graph between the different states. A node corresponds to a state with the mean 
time spent in this state. Each arrow represents the probability of transtion between states.
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# estimation  
mark <- estimation_Markov(d_JK)

plot(mark)

}
\author{
Cristian Preda
}
