% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{plot_Markov}
\alias{plot_Markov}
\title{Plot the transition graph}
\usage{
plot_Markov(res_Markov)
}
\arguments{
\item{res_Markov}{output of \code{\link{estimation_Markov}} function}
}
\description{
Plot the transition graph between the different states. A node corresponds to a state with the mean 
time spent in this state. Each arrow represents the probability of transtion between states.
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# estimation  
mark <- estimation_Markov(d_JK)

plot_Markov(mark)

}
\author{
Cristian Preda
}
