% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{plot_pt}
\alias{plot_pt}
\title{Plot probabilities}
\usage{
plot_pt(pt, ...)
}
\arguments{
\item{pt}{output of \code{\link{estimate_pt}}}

\item{...}{extra \code{plot} parameters}
}
\description{
Plot the probabilities of each state at each given time
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

d_JK2 <- msm2msmTmax(d_JK, 10)

pt <- estimate_pt(d_JK2)

plot_pt(pt)

}
\author{
Cristian Preda
}
