% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmca.print.R
\name{print.fmca}
\alias{print.fmca}
\title{Print a \code{fmca} object}
\usage{
\method{print}{fmca}(x, n = 6, ...)
}
\arguments{
\item{x}{\code{fmca} object (see \link{compute_optimal_encoding} function)}

\item{n}{maximal number of rows and cols to print}

\item{...}{Not used.}
}
\value{
No return value, called for side effects
}
\description{
Print a \code{fmca} object
}
\seealso{
Other encoding functions: 
\code{\link{compute_optimal_encoding}()},
\code{\link{get_encoding}()},
\code{\link{plot.fmca}()},
\code{\link{plotComponent}()},
\code{\link{plotEigenvalues}()},
\code{\link{predict.fmca}()},
\code{\link{summary.fmca}()}
}
\concept{encoding functions}
