% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{statetable}
\alias{statetable}
\title{Table of transitions}
\usage{
statetable(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, \code{time} and \code{state} (see \code{\link{generate_Markov_cfd}})}
}
\value{
a vector of length \code{K} containing the total time spent in each state
}
\description{
Calculates a frequency table counting the number of times each pair of states were observed in successive observation times.
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# table of transitions
statetable(d_JK)


}
