% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmca.print.R
\name{summary.fmca}
\alias{summary.fmca}
\title{Object Summaries

Summary of a \code{fmca} object}
\usage{
\method{summary}{fmca}(object, n = 6, ...)
}
\arguments{
\item{object}{\code{fmca} object (see \link{compute_optimal_encoding} function)}

\item{n}{maximal number of rows and cols to print}

\item{...}{Not used.}
}
\value{
No return value, called for side effects
}
\description{
Object Summaries

Summary of a \code{fmca} object
}
\seealso{
\link{compute_optimal_encoding} \link{print.fmca}
}
