% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{boxplot.timeSpent}
\alias{boxplot.timeSpent}
\title{Boxplot of time spent in each state}
\usage{
\method{boxplot}{timeSpent}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{compute_Time_Spent}} function}

\item{...}{not used}
}
\description{
Boxplot of time spent in each state
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# cut at Tmax = 8
d_JK2 <- msm2msmTmax(d_JK, Tmax = 8)

# compute time spent by each id in each state
timeSpent <- compute_Time_Spent(d_JK2)

# plot the result
boxplot(timeSpent)

}
\author{
Quentin Grimonprez
}
