% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfda-package.R
\docType{package}
\name{cfda-package}
\alias{cfda-package}
\title{Categorical Functional Data Analysis}
\description{
Categorical Functional Data Analysis
}
\details{
See the vignette for a detailled example and mathematical background:
\code{RShowDoc("cfda", package = "cfda")}
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
Tmax <- 6
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = Tmax)
d_JK2 <- msm2msmTmax(d_JK, Tmax)

# create basis object
m <- 10
b <- create.bspline.basis(c(0, Tmax), nbasis = m, norder = 4)

# compute encoding
encoding <- compute_optimal_encoding(d_JK2, b, nCores = 1)

# plot eigenvalues
plotEigenvalues(encoding, cumulative = TRUE, normalize = TRUE)

# plot the two first components
plotComponent(encoding, comp = c(1, 2))

# plot the encoding using the first harmonic
plot(encoding)

# extract the encoding using the first harmonic
getEncoding(encoding)


}
\seealso{
\link{compute_optimal_encoding}
}
\keyword{package}
