% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfda-package.R
\docType{package}
\name{cfda-package}
\alias{cfda-package}
\title{Categorical Functional Data Analysis}
\description{
Categorical Functional Data Analysis
}
\examples{
# See the vignette for a detailled example and mathematical background
RShowDoc("cfda", package = "cfda")

}
\keyword{package}
