% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{compute_Duration}
\alias{compute_Duration}
\title{Compute duration of individuals}
\usage{
compute_Duration(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}
}
\value{
a vector containing the duration of each trajectories
}
\description{
For each individual, compute the duration
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)


# compute duration of each individual
duration <- compute_Duration(d_JK)


}
\author{
Cristian Preda, Quentin Grimonprez
}
