% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{compute_number_jumps}
\alias{compute_number_jumps}
\title{Compute the number of jumps}
\usage{
compute_number_jumps(data, countDuplicated = TRUE)
}
\arguments{
\item{data}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}

\item{countDuplicated}{if \code{TRUE}, jumps in the same state are counted as jump}
}
\value{
A vector containing the number of jumps for each individual
}
\description{
For each individual, compute the number of jumps performed
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# compute the number of jumps
nJump <- compute_number_jumps(d_JK)

}
\seealso{
\link{hist.njump}
}
\author{
Cristian Preda, Quentin Grimonprez
}
