% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{compute_optimal_encoding}
\alias{compute_optimal_encoding}
\title{Compute the optimal encodings for each state}
\usage{
compute_optimal_encoding(data_msm, basisobj, nCores = max(1,
  ceiling(detectCores()/2)), ...)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state (integer starting at 1). All individual must end at the same time Tmax (use \code{\link{msm2msmTmax}}).}

\item{basisobj}{is a basis create using \code{fda} package.}

\item{nCores}{number of Cores used for parallelization. Default is the half of cores.}

\item{...}{parameters for \code{\link{integrate}} function.}
}
\value{
A list containing:
\itemize{
  \item eigenvalues eigenvalues
  \item alpha optimal encoding coefficients
  \item pc principal components
  \item F matrix containing the \eqn{F_{(x,i)(y,j)}}
  \item V matrix containing the \eqn{V_{(x,i)}}
  \item G covariance matrix of \code{V}
  \item basisobj \code{basisobj} parameter
}
}
\description{
Compute the optimal encodings for each state
}
\details{
see the vignette for the mathematical background: \code{vignette("cfda", package = "cfda")}
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
Tmax <- 10
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = Tmax)
d_JK2 <- msm2msmTmax(d_JK, 10)

# create basis object
m <- 10
b <- create.bspline.basis(c(0, Tmax), nbasis = m, norder = 4)

# compute encoding
encoding <- compute_optimal_encoding(d_JK2, b, nCores = 2)

# plot the optimal encoding
plot(encoding)

# plot the two first components
plotComponent(encoding, comp = c(1, 2))

}
\references{
Deville J.C. (1982) Analyse de donn\'ees chronologiques qualitatives: comment analyser des calendriers ?, Annales de l'INSEE, No 45, p. 45-104.
}
\seealso{
\link{plot.fmca} \link{plotComponent}
}
\author{
Cristian Preda
}
