% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretreatment.R
\name{cut_data}
\alias{cut_data}
\title{Cut data to a maximal given time}
\usage{
cut_data(data_msm, Tmax)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}

\item{Tmax}{max time considered}
}
\value{
a data.frame with the same format as \code{data_msm} where each individual has \code{Tmax} as last time entry.
}
\description{
Cut data to a maximal given time
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK = generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)
tail(d_JK)

# cut at Tmax = 8
d_JK2 <- cut_data(d_JK, Tmax = 8)
tail(d_JK2)

}
\author{
Cristian Preda
}
