% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{estimate_pt}
\alias{estimate_pt}
\title{Estimate probabilities to be in each state}
\usage{
estimate_pt(data_msm)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state. All individual must end at the same time Tmax (use \code{\link{msm2msmTmax}}).}
}
\value{
A list of two elements:
\itemize{
  \item{t: vector of time}
  \item{pt: a matrix with K (= number of states) rows and with \code{length(t)} columns containing the probabilities to be in each state at each time.}
}
}
\description{
Estimate probabilities to be in each state
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

d_JK2 <- msm2msmTmax(d_JK, 10)

# estimate probabilities
estimate_pt(d_JK2)

}
\seealso{
\link{plot.pt}
}
\author{
Cristian Preda, Quentin Grimonprez
}
