% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation_Markov}
\alias{estimation_Markov}
\title{Estimate transition matrix and spent time}
\usage{
estimation_Markov(data)
}
\arguments{
\item{data}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}
}
\value{
list of two elements: \code{Q}, the estimated transition matrix, and \code{lambda}, the estimated time spent in each state
}
\description{
Calculates crude initial values for transition intensities by assuming that the data represent the exact transition times of the Markov process.
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# estimation  
mark <- estimation_Markov(d_JK)
mark$Q
mark$lambda

}
\seealso{
\link{plot.Markov}
}
\author{
Cristian Preda
}
