% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generate_2State}
\alias{generate_2State}
\title{Generate data following a 2 states model}
\usage{
generate_2State(n)
}
\arguments{
\item{n}{number of individuals}
}
\value{
a data.frame with 3 columns: \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, new state.
}
\description{
Generate individuals such that each individual starts at time 0 with state 0 and then an unique change to state 1 occurs at a time \eqn{t} generated using an uniform law between 0 and 1.
}
\author{
Cristian Preda, Quentin Grimonprez
}
