% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{get_state}
\alias{get_state}
\title{Extract the state of each individual at a given time}
\usage{
get_state(data_msm, t)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state (integer starting at 1).}

\item{t}{real}
}
\value{
a vector containing the state of each individual at time t
}
\description{
Extract the state of each individual at a given time
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# get the state of each individuals at time t = 6
get_state(d_JK, 6)

}
\author{
Cristian Preda
}
