% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{plot.fmca}
\alias{plot.fmca}
\title{Plot the optimal encoding}
\usage{
\method{plot}{fmca}(x, harm = 1, col = NULL, ...)
}
\arguments{
\item{x}{output of \code{\link{compute_optimal_encoding}} function}

\item{harm}{harmonic to use for the encoding}

\item{col}{a vector containing color for each state}

\item{...}{not used}
}
\description{
Plot the optimal encoding
}
\details{
The encoding is \eqn{a_{x} \approx \sum_{i=1}^m \alpha_{x,i}\phi_i}.
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
Tmax <- 6
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = Tmax)
d_JK2 <- msm2msmTmax(d_JK, Tmax)

# create basis object
m <- 10
b <- create.bspline.basis(c(0, Tmax), nbasis = m, norder = 4)

# compute encoding 
encoding <- compute_optimal_encoding(d_JK2, b, nCores = 1)

# plot the encoding using 1 harmonic
plot(encoding)

}
\seealso{
\link{plotComponent} \link{plotEigenvalues}
}
\author{
Quentin Grimonprez
}
