% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{plot.Markov}
\alias{plot.Markov}
\title{Plot the transition graph}
\usage{
\method{plot}{Markov}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{estimate_Markov}} function}

\item{...}{parameters of \code{plotmat} function from \code{diagram} package (see details).}
}
\description{
Plot the transition graph between the different states. A node corresponds to a state with the mean 
time spent in this state. Each arrow represents the probability of transtion between states.
}
\details{
Some usefull extra parameters:
\itemize{
  \item \code{dtext} controls the position of arrow text relative to arrowhead (default = 0.3).
  \item \code{relsize}	scaling factor for size of the graph (default = 1).
  \item \code{box.size}	size of label box, one value or a vector with dimension = number of rows of \code{x$Q}.
  \item \code{box.cex}	relative size of text in boxes, one value or a vector with dimension=number of rows of \code{x$Q}.
}
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 100, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# estimation  
mark <- estimate_Markov(d_JK)

# transition graph
plot(mark)

}
\author{
Cristian Preda
}
