% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{plot.pt}
\alias{plot.pt}
\title{Plot probabilities}
\usage{
\method{plot}{pt}(x, ribbon = FALSE, ...)
}
\arguments{
\item{x}{output of \code{\link{estimate_pt}}}

\item{ribbon}{if TRUE, use ribbon to plot probabilities}

\item{...}{only if \code{ribbon = TRUE}, parameter \code{addBorder}, if TRUE, add black border to the ribbons.}
}
\description{
Plot the probabilities of each state at each given time
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

d_JK2 <- msm2msmTmax(d_JK, 10)

pt <- estimate_pt(d_JK2)

plot(pt, ribbon = TRUE)

}
\author{
Quentin Grimonprez
}
