% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{plotData}
\alias{plotData}
\title{Plot categorical functional data}
\usage{
plotData(data_msm, addLabel = TRUE, addBorder = TRUE)
}
\arguments{
\item{data_msm}{data.frame containing \code{id}, id of the trajectory, \code{time}, time at which a change occurs and \code{state}, associated state.}

\item{addLabel}{If TRUE, add id labels}

\item{addBorder}{If TRUE add black border to each individuals}
}
\value{
each row represent an individual over [0:Tmax]. The color at a given time gives the state of the individual.
}
\description{
Plot categorical functional data
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov_cfd(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = 10)

# add a line with time Tmax at the end of each individuals
d_JKT <- msm2msmTmax(d_JK, Tmax = 10)

plotData(d_JKT)
  
}
\author{
Cristian Preda, Quentin Grimonprez
}
