% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{plotEigenvalues}
\alias{plotEigenvalues}
\title{Plot Eigenvalues}
\usage{
plotEigenvalues(x, cumulative = FALSE, normalize = FALSE, ...)
}
\arguments{
\item{x}{output of \code{\link{compute_optimal_encoding}} function}

\item{cumulative}{if TRUE, plot the cumualtive eigenvalues}

\item{normalize}{if TRUE eigenvalues are normalized for summing to 1}

\item{...}{\code{geom_point} parameters}
}
\value{
a \code{ggplot} object that can be modified using \code{ggplot2} package.
}
\description{
Plot Eigenvalues
}
\examples{
# simulate the Jukes Cantor models of nucleotides replacement. 
K <- 4
Tmax <- 6
QJK <- matrix(1/3, nrow = K, ncol = K) - diag(rep(1/3, K))
lambda_QJK <- c(1, 1, 1, 1)
d_JK <- generate_Markov(n = 10, K = K, Q = QJK, lambda = lambda_QJK, Tmax = Tmax)
d_JK2 <- cut_data(d_JK, Tmax)

# create basis object
m <- 10
b <- create.bspline.basis(c(0, Tmax), nbasis = m, norder = 4)

# compute encoding
encoding <- compute_optimal_encoding(d_JK2, b, nCores = 1)

# plot eigenvalues
plotEigenvalues(encoding, cumulative = TRUE, normalize = TRUE)

# modify the plot using ggplot2
library(ggplot2)
plotEigenvalues(encoding, shape = 23) +
   labs(caption = "Jukes Cantor models of nucleotides replacement")

}
\seealso{
\link{plot.fmca} \link{plotComponent}
}
\author{
Quentin Grimonprez
}
