% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmca.print.R
\name{print.fmca}
\alias{print.fmca}
\title{Print \code{fmca} object

Print a \code{fmca} object}
\usage{
\method{print}{fmca}(x, n = 6, ...)
}
\arguments{
\item{x}{\code{fmca} object (see \link{compute_optimal_encoding} function)}

\item{n}{maximal number of rows and cols to print}

\item{...}{Not used.}
}
\description{
Print \code{fmca} object

Print a \code{fmca} object
}
\seealso{
\link{compute_optimal_encoding}
}
