% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmca.print.R
\name{summary.fmca}
\alias{summary.fmca}
\title{Object Summaries}
\usage{
\method{summary}{fmca}(object, n = 6, ...)
}
\arguments{
\item{object}{\code{fmca} object (see \link{compute_optimal_encoding} function)}

\item{n}{maximal number of rows and cols to print}

\item{...}{Not used.}
}
\value{
No return value, called for side effects
}
\description{
Summary of a \code{fmca} object
}
\seealso{
Other encoding functions: 
\code{\link{compute_optimal_encoding}()},
\code{\link{get_encoding}()},
\code{\link{plot.fmca}()},
\code{\link{plotComponent}()},
\code{\link{plotEigenvalues}()},
\code{\link{predict.fmca}()},
\code{\link{print.fmca}()}
}
\concept{encoding functions}
